UPDATE: ai control are in a separated file
