+++ index.html
@@ -90,6 +90,7 @@ option {
         <button id="startButton">Start Game</button>
     </div>
     <button id="pauseButton">Pause</button>
-    <script src="pong.js" type="text/javascript"></script>
+    <script src="pong.js" type="module"></script>
+    <script src="pongAI.js" type="module"></script>
 </body>
 </html>
